+++ crates/testing/src/task_helpers.rs
@@ -180,7 +180,7 @@ pub fn vid_init<TYPES: NodeType>(
     // calculate the last power of two
     // TODO change after https://github.com/EspressoSystems/jellyfish/issues/339
     // issue: https://github.com/EspressoSystems/HotShot/issues/2152
-    let chunk_size = ((num_committee as u64).next_power_of_two() / 2) as usize;
+    let chunk_size = num_committee.next_power_of_two() / 2;
 
     // TODO <https://github.com/EspressoSystems/HotShot/issues/1686>
     let srs = hotshot_types::data::test_srs(num_committee);
